fixed lost layout after changeing the headlines to display
git-svn-id: http://svn.egroupware.org/egroupware/trunk@11521 923e3c2e-ae13-0410-ba18-ced7211e33de

+++ headlines/preferences.php
@@ -28,7 +28,10 @@
 			{
 				foreach($GLOBALS['phpgw_info']['user']['preferences']['headlines'] as $n => $name)
 				{
-					$GLOBALS['phpgw']->preferences->delete('headlines',$n);
+					if ($n != 'headlines_layout')
+					{
+						$GLOBALS['phpgw']->preferences->delete('headlines',$n);
+					}
 				}
 			}
